+++ lib/kms_encrypted/log_subscriber.rb
@@ -1,0 +1,17 @@
+module KmsEncrypted
+  class LogSubscriber < ActiveSupport::LogSubscriber
+    def decrypt(event)
+      return unless logger.debug?
+
+      name = "Decrypt (#{event.duration.round(1)}ms)"
+      debug "  #{color(name, YELLOW, true)}  Context: #{event.payload[:context].inspect}"
+    end
+
+    def generate_data_key(event)
+      return unless logger.debug?
+
+      name = "Generate Data Key (#{event.duration.round(1)}ms)"
+      debug "  #{color(name, YELLOW, true)}  Context: #{event.payload[:context].inspect}"
+    end
+  end
+end