+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `kms_keys` method to models
 - Reset data keys when record is reloaded
 - Added `kms_client`
+- Added ActiveSupport notifications
 
 ## 0.1.3